New folder/requirements.txt
streamlit
google-generativeai
pandas
numpy

To install the packages listed in the `requirements.txt` file, use the following command:

```sh
pip install -r /R:/New\ folder/New\ folder/streamlit-app/requirements.txt
```

To run the Streamlit app, use the following command:

```sh
streamlit run app.py
```

To create a virtual environment, use the following commands:

```sh
python -m venv env
source env/bin/activate  # On Windows use `env\Scripts\activate`
```